requirements.txt
python-telegram-bot[rate-limiter]==21.5
pymongo==4.8.0
openai==1.43.0
tiktoken>=0.7.0
PyYAML==6.0.2
requests==2.32.3
